removed dependency to gdiplus
When tinkering with QT4 (actually with coverity scan ;-) I did only want to use cl.exe instead of qtmake. Doing so I've just took the sources, and checked for necessary options to find out that SVGThumbnailExtension.pro has three libraries listed and and only 2 where necessary (despite of the qt libraries).
If it is not known to be necessary for some environments (and the PR passes the build test) it should be removed.